Remove party_field_lookup button that prevents submitting the edit-prod form with the enter key
The party lookup widget incorporates a submit button, hidden with JS, that forces a re-render of the form with the party ID filled in - this allows users with JS disabled to fill in the field.

4fa87c557c67fdc0438d796f769c33293f9ded25 fixed the case where this wasn't being respected on leaving the field blank, but caused a bigger issue: submitting the form with the enter key will trigger the first submit button in the HTML, which in this case is the hidden party_field_lookup button.

The solution is to remove that button completely in Javascript, rather than just hiding it - which matches what already happens with the byline field. Non-JS users will still have to explicitly click the submit button when they're done, but sub-optimal UI for non-JS users is not exactly news, and the fact that it works at all in 2020 is better than 99% of the internet. :-P