Recognise (some) Mastodon URLs
Fixes #544. A quick check of the database revealed that there are currently 14 Mastodon account links across 9 different instances, so I've hard-coded those 9. If we ever reach the point that managing that list manually is unsustainable _then_ we can consider building a fancy polling mechanism to discover Mastodon instances, but right now that would be massive over-engineering. :-)