Reject youtube video IDs with invalid characters
If the `v` parameter of a regular Youtube link contains multiple slashes (for example, someone has pasted another URL after it) this will dutifully be saved to the database as the parameter. However, since we represent timestamped links as `<video_id>/<timestamp>`, unpacking it again (to output the URL) will lead to it trying to split on slashes, and that code doesn't account for the possibility of more than one slash. Fix this by properly rejecting video IDs with invalid characters in them.