Add test coverage for meteoriks candidates export
Exclude a whole load of lines, because most of them are one-off special cases for weird real-world data that would be pointless to replicate in tests. It's not like we need to keep this export continually in working condition anyhow, since I'm only going to run it once a year (and when I do there's inevitably a bunch of data / logic fixing needed for it to succeed).