Handle alphanumeric keys in keydown handler rather than keypress
Current versions of Firefox do not have the behaviour we rely on where setting focus to an input inside a keypress event, and leaving the event uncancelled (by returning true), causes the input to receive the keypress (and thus treat it as the first character of the input). To get around this, we do the refocusing in the keydown event, so that the focus is already on the input when the keypress event fires.